fix: noticeDetail페이지 로직 리팩토링 완료
